# edwards: improve the performance of Add, MixedAdd and IsOnCurve