… ecc/bls12-381/bandersnatch/point_test.go
… ecc/bls12-381/twistededwards/point_test.go
… ecc/bls24-315/twistededwards/point_test.go
… ecc/bls24-317/twistededwards/point_test.go
… ecc/bn254/twistededwards/point_test.go
… ecc/bw6-633/twistededwards/point_test.go
… ecc/bw6-756/twistededwards/point_test.go
… ecc/bw6-761/twistededwards/point_test.go
… internal/generator/edwards/template/tests/point.go.tmpl